fix: better support of exotic features of git config file format. (fixes #4)
git config file format allows ambiguous keys:

    [a "b.c"]
        d = foo
    [a.b "c"]
        e = foo
    [a.b.c]
        f = foo

Are all valid. So code was simplified to use directly ``git config``.
This simplification will deal also with cases where section could be
attributed values:

    [a "b"]
        c = foo
    [a]
        b = foo

By avoiding to parse the entire content of the file, and relying on
``git config`` implementation we ensure to remain compatible and not
re-implement the parsing of this file format.