chg: dev: moved commit's attribute retrieval out of ``__init__`` towards ``__getattr__``.
This allows to fetch attribute on demand. This won't have any performance hit yet as
all GitCommit will trigger the full fetch at least once for now.

But we now can initialize the GitCommit attributes with other means. And if we don't use
an attribute that wasn't computed, we will spare a ``git show`` call.