fix: nasty hidden bug that would break python3 (fixes #27)
Actually this bug was revealed by python3 random hashes (thanks to
@rschoon for the hint) and could be reproduced on python2.7 with ``-R``
mode.

The ``git show`` command actually will behave differently if given a tag
reference and print random unexpected information before using the
format string. This would prefix a lot of mess to the first field being
asked in the format string.

And this first field is dependent on the internal order of a dict, and
this order is not important as such, and so nothing was done on this
part.

On python2.7, somehow, it would always be the same order that revealed
to have no consequence (probably one of the rare field not used in
current changelogs).

Python3 or Python2.7 -R would shuffle this order and then trigger the
error whenever this prefix would be appended to actually important
fields that went into some further processing (as casted to int for
the timestamp for instance).