fix: output warning on stderr in some weird cases (fixes #52)
If no tag are found in the repository, or no tag matches the filter
regex, or if all commits are ignored... this will lead to disturbing but
legit outputs from ``gitchangelog``. So as to help diagnose what is
going on, additional warnings are then printed when edge cases are
encountered.