fix: fixed encoding issue when reading UTF-8 git logs with a different default locale.
Windows platform were more likely to get hit by this bug as their
default code page is not ``utf-8``. It was fixed by using an explicit
encoding when reading git logs. The default value for this encoding
can now be set in the ``gitchangelog``'s config file, per-repository.
Although, this option should be only set in pathological configuration
as the default behavior is to use ``git config i18n.logOutputEncoding``
when set, or if not set, ``utf-8``, which is the default log encoding
of git.