new: ``-p`` command line option to limit commits to a specified path
Using this option one can limit the changelog to only those commits that
affect the specified path. This is similar to passing a path option to
``git log``. It is very useful in cases where a repository has multiple
modules and one wishes to generate a change log for only one module or
directory. For example a repository with a main application and multiple
supporting utilities.